+++ src/sql-parser/src/ast/defs/value.rs
@@ -18,7 +18,160 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::ast::display::{AstDisplay, AstFormatter};
+use std::fmt;
+
+use repr::adt::datetime::DateTimeField;
+
+use crate::ast::display::{self, AstDisplay, AstFormatter};
+
+#[derive(Debug)]
+pub struct ValueError(String);
+
+impl std::error::Error for ValueError {}
+
+impl fmt::Display for ValueError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+/// Primitive SQL values such as number and string
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum Value {
+    /// Numeric literal
+    Number(String),
+    /// 'string value'
+    SingleQuotedString(String),
+    /// X'hex value'
+    HexStringLiteral(String),
+    /// Boolean value true or false
+    Boolean(bool),
+    /// INTERVAL literals, roughly in the following format:
+    ///
+    /// ```text
+    /// INTERVAL '<value>' <leading_field> [ TO <last_field>
+    ///     [ (<fractional_seconds_precision>) ] ]
+    /// ```
+    /// e.g. `INTERVAL '123:45.678' MINUTE TO SECOND(2)`.
+    Interval(IntervalValue),
+    /// `NULL` value
+    Null,
+}
+
+impl AstDisplay for Value {
+    fn fmt(&self, f: &mut AstFormatter) {
+        match self {
+            Value::Number(v) => f.write_str(v),
+            Value::SingleQuotedString(v) => {
+                f.write_str("'");
+                f.write_node(&display::escape_single_quote_string(v));
+                f.write_str("'");
+            }
+            Value::HexStringLiteral(v) => {
+                f.write_str("X'");
+                f.write_str(v);
+                f.write_str("'");
+            }
+            Value::Boolean(v) => f.write_str(v),
+            Value::Interval(IntervalValue {
+                value,
+                precision_high: _,
+                precision_low: _,
+                fsec_max_precision: Some(fsec_max_precision),
+            }) => {
+                f.write_str("INTERVAL '");
+                f.write_node(&display::escape_single_quote_string(value));
+                f.write_str("' SECOND (");
+                f.write_str(fsec_max_precision);
+                f.write_str(")");
+            }
+            Value::Interval(IntervalValue {
+                value,
+                precision_high,
+                precision_low,
+                fsec_max_precision,
+            }) => {
+                f.write_str("INTERVAL '");
+                f.write_node(&display::escape_single_quote_string(value));
+                f.write_str("'");
+                match (precision_high, precision_low, fsec_max_precision) {
+                    (DateTimeField::Year, DateTimeField::Second, None) => {}
+                    (DateTimeField::Year, DateTimeField::Second, Some(ns)) => {
+                        f.write_str(" SECOND(");
+                        f.write_str(ns);
+                        f.write_str(")");
+                    }
+                    (DateTimeField::Year, low, None) => {
+                        f.write_str(" ");
+                        f.write_str(low);
+                    }
+                    (high, low, None) => {
+                        f.write_str(" ");
+                        f.write_str(high);
+                        f.write_str(" TO ");
+                        f.write_str(low);
+                    }
+                    (high, low, Some(ns)) => {
+                        f.write_str(" ");
+                        f.write_str(high);
+                        f.write_str(" TO ");
+                        f.write_str(low);
+                        f.write_str("(");
+                        f.write_str(ns);
+                        f.write_str(")");
+                    }
+                }
+            }
+            Value::Null => f.write_str("NULL"),
+        }
+    }
+}
+impl_display!(Value);
+
+/// An intermediate value for Intervals, which tracks all data from
+/// the user, as well as the computed ParsedDateTime.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct IntervalValue {
+    /// The raw `[value]` that was present in `INTERVAL '[value]'`
+    pub value: String,
+    /// The most significant DateTimeField to propagate to Interval in
+    /// compute_interval.
+    pub precision_high: DateTimeField,
+    /// The least significant DateTimeField to propagate to Interval in
+    /// compute_interval.
+    /// precision_low is also used to provide a TimeUnit if the final
+    /// part of `value` is ambiguous, e.g. INTERVAL '1-2 3' DAY uses
+    /// 'day' as the TimeUnit for 3.
+    pub precision_low: DateTimeField,
+    /// Maximum nanosecond precision can be specified in SQL source as
+    /// `INTERVAL '__' SECOND(_)`.
+    pub fsec_max_precision: Option<u64>,
+}
+
+impl Default for IntervalValue {
+    fn default() -> Self {
+        Self {
+            value: String::default(),
+            precision_high: DateTimeField::Year,
+            precision_low: DateTimeField::Second,
+            fsec_max_precision: None,
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn iterate_datetimefield() {
+        use DateTimeField::*;
+        assert_eq!(
+            Year.into_iter().take(10).collect::<Vec<_>>(),
+            vec![Month, Day, Hour, Minute, Second]
+        )
+    }
+}
 
 /// SQL data types
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
