+++ src/sql-parser/tests/testdata/error
@@ -91,15 +91,6 @@ SELECT count(ALL DISTINCT + x) FROM customer
 Cannot specify both ALL and DISTINCT in function: count
 
 parse-statement
-SELECT EXTRACT(MILLISECOND FROM d)
-----
-error:
-Parse error:
-SELECT EXTRACT(MILLISECOND FROM d)
-               ^^^^^^^^^^^
-Expected valid extract field, found: MILLISECOND
-
-parse-statement
 SELECT INTERVAL '01:01.01' MINUTE (5) TO SECOND (5)
 ----
 error:
